refactoring such that ML model related stuff is
out of the openpmd-streaming-continual-learning.py
and put into a separate directory
