+++ tools/models/__init__.py
@@ -1,0 +1,18 @@
+"""
+Model-related modules for openpmd-learning.
+
+This package contains:
+- Model architectures
+- Model initialization and loading functions
+- Configuration utilities
+"""
+
+from models.architectures import ModelFinal
+from models.model_factory import load_objects, get_VAE_encoder_kwargs, get_VAE_decoder_kwargs
+
+__all__ = [
+    'ModelFinal',
+    'load_objects',
+    'get_VAE_encoder_kwargs',
+    'get_VAE_decoder_kwargs',
+]
